# openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "InvitationInput": {
        "additionalProperties": false,
        "properties": {
          "password": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "type": "object"
      },
...
      "PasswordInput": {
        "additionalProperties": false,
        "properties": {
          "current_password": {
            "type": "string"
          },
          "new_password": {
            "type": "string"
          }
        },
        "required": [
          "current_password",
          "new_password"
        ],
        "type": "object"
      },
...
      "ProblemDetails": {
        "properties": {
          "detail": {
            "type": "string"
          },
          "status": {
            "format": "int32",
            "maximum": 599,
            "minimum": 100,
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "format": "uri-reference",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail"
        ],
        "type": "object"
      },
...
      "SessionCreatedResponse": {
        "properties": {
          "csrf_token": {
            "type": "string"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "user": {
            "$ref": "#/components/schemas/UserResource"
          }
        },
        "required": [
          "user",
          "csrf_token",
          "permissions"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "scheme": "bearer",
        "type": "http"
      },
      "sessionCookie": {
        "in": "cookie",
        "name": "racebin_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {
      "name": "Racebin contributors"
    },
    "description": "API-first, self-hosted paste bin and file share.",
    "license": {
      "identifier": "MIT",
      "name": "MIT"
    },
    "title": "Racebin API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/account/password": {
      "patch": {
        "operationId": "change_password",
        "parameters": [
          {
            "description": "Current browser session CSRF token",
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Password changed and existing sessions revoked"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Invalid password or browser session required"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Current password is incorrect"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "CSRF validation failed"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "tags": [
          "account"
        ]
      }
    },
...
    "/invitations/{token}/redeem": {
      "post": {
        "operationId": "redeem_invitation",
        "parameters": [
          {
            "description": "One-time invitation token",
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvitationInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionCreatedResponse"
                }
              }
            },
            "description": "Account and browser session created",
            "headers": {
              "Set-Cookie": {
                "description": "HTTP-only Racebin session cookie",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Invalid invitation, username, or password"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Username already exists"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Invitation-attempt rate limit exceeded",
            "headers": {
              "Retry-After": {
                "description": "Seconds before another attempt",
                "schema": {
                  "format": "int64",
                  "type": "integer"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal error"
          }
        },
        "security": [
          {}
        ],
        "tags": [
          "account"
        ]
      }
    },
...
    "/password-resets/{token}": {
      "post": {
        "operationId": "reset_password",
        "parameters": [
          {
            "description": "One-time password-reset token",
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Password reset and existing sessions revoked"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Invalid, expired, or consumed token; or invalid password"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Reset-attempt rate limit exceeded",
            "headers": {
              "Retry-After": {
                "description": "Seconds before another attempt",
                "schema": {
                  "format": "int64",
                  "type": "integer"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal error"
          }
        },
        "security": [
          {}
        ],
        "tags": [
          "account"
        ]
      }
    },
...
    "/session": {
      "delete": {
        "operationId": "logout",
        "parameters": [
          {
            "description": "Current browser session CSRF token",
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Session deleted",
            "headers": {
              "Set-Cookie": {
                "description": "Expired Racebin session cookie",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Browser session required"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "CSRF validation failed"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "tags": [
          "account"
        ]
      },
      "get": {
        "operationId": "get_session",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Current browser session, bearer-key identity, or anonymous state"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Invalid bearer credential"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal error"
          }
        },
        "security": [
          {},
          {
            "bearerAuth": []
          },
          {
            "sessionCookie": []
          }
        ],
        "tags": [
          "account"
        ]
      },
      "post": {
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionCreatedResponse"
                }
              }
            },
            "description": "Browser session created",
            "headers": {
              "Set-Cookie": {
                "description": "HTTP-only Racebin session cookie",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Malformed credentials"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Invalid credentials"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Login rate limit exceeded",
            "headers": {
              "Retry-After": {
                "description": "Seconds before another attempt",
                "schema": {
                  "format": "int64",
                  "type": "integer"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal error"
          }
        },
        "security": [
          {}
        ],
        "tags": [
          "account"
        ]
      }
    },
    "/session/reauthenticate": {
      "post": {
        "operationId": "reauthenticate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReauthenticateInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Session authorized for sensitive owner operations for ten minutes"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Password is incorrect"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Browser session and CSRF token required"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "tags": [
          "account"
        ]
      }
    }
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
...
}